feat(vulkan): implement wrapExternalImageResource (11/11 IRenderDevice)
Widen ImageResourceHandle.value to 64-bit (GL texture IDs still fit; all GL uses static_cast<GLuint>) so it can carry a foreign VkImage. wrapExternalImageResource now creates a non-owning VulkanTextureResource over the borrowed VkImage (owns only view+sampler; never destroys the image). Adds a Vulkan-specific nativeImageHandle accessor to round-trip an owned texture's VkImage. New VulkanExternalImageTests. Completes IRenderDevice parity (11/11).

+++ src/render/RenderTypes.h
@@ -10,7 +10,7 @@ class RenderContext;
 struct ScissorState;
 
 struct ImageResourceHandle {
-    std::uint32_t value = 0;
+    std::uint64_t value = 0;
 
     bool isValid() const { return value != 0; }
 };
